Refactored email notifier plugin to be in correct package and updated javadoc plugin settings to disable doclint
