+++ nucleus/pom.xml
@@ -338,7 +338,12 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-javadoc-plugin</artifactId>
-                    <version>2.9.1</version>
+                    <version>2.10.4</version>
+                    <configuration>
+                        <detectOfflineLinks>false</detectOfflineLinks>
+                        <skip>${javadoc.skip}</skip>
+                        <additionalparam>-Xdoclint:none</additionalparam>
+                    </configuration>
                 </plugin>
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
@@ -694,14 +699,10 @@
                 <configuration>
                     <isFailureFatal>${command.security.maven.plugin.isFailureFatal}</isFailureFatal>
                 </configuration>
-            </plugin>
+            </plugin> 
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-javadoc-plugin</artifactId>
-                <configuration>
-                    <detectOfflineLinks>false</detectOfflineLinks>
-                    <skip>${javadoc.skip}</skip>
-                </configuration>
             </plugin>
         </plugins>
     </build>
@@ -1088,11 +1089,6 @@
                 <groupId>jline</groupId>
                 <artifactId>jline</artifactId>
                 <version>2.14.2</version>
-            </dependency>
-            <dependency>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-javadoc-plugin</artifactId>
-                <version>2.8</version>
             </dependency>
             <dependency>
                 <groupId>org.fusesource.hawtjni</groupId>
